Test Cases, pdf showing all the research and information I've gathered (so far)
Test cases showing the amazing improvement when using the libdfp vs libmpdec. Making mpdec work like that through will provide a significant increase in speed.
Following that, additional steps will be needed in the context of Python String<->float/int Conversion Optimization for IBM Z
The PDF is the collection of all the information I've gathered in the most concise way I could write it (but without removing important information).